pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "wry_bokeh_helper"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = ["pillow"]


[project.optional-dependencies]
tests = ["pytest"]
dev = ["ruff"]

[tool.maturin]
features = ["pyo3/extension-module"]
